requirements.txt
pybind11
moderngl
numpy
PyGLM
pyGLFW
Pillow
git+https://github.com/pyimgui/pyimgui.git@dev/version-2.0
